Matt effected population trigger

+++ data/triggers.sql
@@ -1,3 +1,4 @@
+  
 DROP TRIGGER IF EXISTS ProduceDamage; 
 
 DELIMITER $$
@@ -13,6 +14,7 @@ DELIMITER ;
 
 
 
+
 DROP TRIGGER IF EXISTS CreateJunction; 
 
 DELIMITER $$
@@ -23,13 +25,13 @@ ON `earthquake` FOR EACH ROW
 BEGIN
 
 	DECLARE radius FLOAT DEFAULT 0.0;
-    DECLARE populationInRadius FLOAT DEFAULT 0;
+  DECLARE populationInRadius FLOAT DEFAULT 0;
     
     SET radius= 2 * NEW.mag * NEW.mag;
 
 	   INSERT INTO earthquake_city
-      SELECT NEW.id,city.id
-		FROM   city
+     SELECT NEW.id,city.id
+		    FROM   city
         WHERE  St_distance_sphere(Point(NEW.longitude, NEW.latitude), Point(
                     city.longitude, city.latitude)) * .000621371192 < radius;
     
@@ -37,6 +39,7 @@ BEGIN
 END $$
     
     DELIMITER ;
+
 
 DROP TRIGGER IF EXISTS FindPopulation; 
 
@@ -47,7 +50,7 @@ BEFORE INSERT
 ON `earthquake` FOR EACH ROW
 BEGIN
 	DECLARE radius FLOAT DEFAULT 0.0;
-    DECLARE populationInRadius INT DEFAULT 0;
+    DECLARE populationInRadius FLOAT DEFAULT 0;
     
     SET radius= 2 * NEW.mag * NEW.mag;
     
@@ -56,11 +59,13 @@ BEGIN
         FROM   city
         WHERE  St_distance_sphere(Point(NEW.longitude, NEW.latitude), Point(
                     city.longitude, city.latitude)) * .000621371192 < radius;  
-   IF(populationInRadius IS NULL) THEN
+                    
+    IF(populationInRadius IS NULL) THEN
             SET populationInRadius = 0;
-        END IF;
-    
+	END IF;
     SET NEW.effected_population=populationInRadius;
+    
+    
 END $$
     
 DELIMITER ;
